requirements.txt
Flask==0.11.1
get
gunicorn==19.6.0
post
setupfiles
requests
wikipedia
line-bot-sdk
pandas==0.20.3

